chore: Add provenance field to My Collection artworks
