+++ src/lib/Components/TextArea.tsx
@@ -1,0 +1,38 @@
+import { color, space, Text } from "palette"
+import React, { useState } from "react"
+import { TextInput, TextInputProps } from "react-native"
+import styled from "styled-components/native"
+
+const StyledTextArea = styled(TextInput)`
+  border: solid 1px;
+  padding: ${space(1)}px;
+  height: 88px;
+`
+
+interface TextAreaProps extends TextInputProps {
+  title: string
+}
+export const TextArea: React.FC<TextAreaProps> = ({ title, ...props }) => {
+  const [borderColor, setBorderColor] = useState(color("black10"))
+
+  return (
+    <>
+      {!!title && (
+        <Text mb={1} variant="mediumText">
+          {title}
+        </Text>
+      )}
+      <StyledTextArea
+        {...props}
+        onFocus={() => {
+          setBorderColor(color("purple100"))
+        }}
+        onBlur={() => {
+          setBorderColor(color("black10"))
+        }}
+        style={{ borderColor }}
+        multiline={true}
+      />
+    </>
+  )
+}
