+++ src/lib/Scenes/MyCollection/State/__tests__/MyCollectionArtworkModel-tests.tsx
@@ -39,6 +39,7 @@ describe("MyCollectionArtworkModel", () => {
               medium: "some-medium",
               metric: "in",
               photos: [fakePhoto("somepath")],
+              provenance: "some-provenance",
               title: "some-title",
               width: "some-width",
             },
@@ -65,6 +66,7 @@ describe("MyCollectionArtworkModel", () => {
       medium: "",
       metric: "",
       photos: [],
+      provenance: "",
       title: "",
       width: "",
     }
